Refactor App component responsibilities
